+++ packages/extension-contracts/src/index.ts
@@ -11,4 +11,5 @@ export * from "./manifest.js";
 export * from "./permissions.js";
 export * from "./settings-contributions.js";
 export * from "./status-bar.js";
+export * from "./topbar.js";
 export * from "./versioning.js";